COR-397: refactor recorder slice, add errors to form
