+++ frontend/apps/pwa/src/app/store.ts
@@ -4,7 +4,7 @@ import formReducer from '../reducers/form';
 import folderReducer from '../reducers/folder';
 import databaseReducer from '../reducers/database';
 import recordsReducer from '../reducers/records';
-import recorderReducer from '../reducers/recorder';
+import recorderSlice from '../reducers/Recorder';
 import formerSlice from '../reducers/Former';
 
 export const store = configureStore({
@@ -13,7 +13,7 @@ export const store = configureStore({
     folders: folderReducer,
     databases: databaseReducer,
     records: recordsReducer,
-    recorder: recorderReducer,
+    recorder: recorderSlice.reducer,
     former: formerSlice.reducer,
   },
   devTools: true,
