+++ frontend/apps/pwa/src/features/recorder/RecordEditor.container.tsx
@@ -8,19 +8,23 @@ import { fetchForms, selectRootForm } from '../../reducers/form';
 import { recordActions } from '../../reducers/records';
 import {
   postRecord,
-  recorderActions,
   resetForm,
+} from '../../reducers/Recorder/recorder.reducers';
+import recorderSlice from '../../reducers/Recorder';
+import {
   selectCurrentForm,
   selectCurrentRecord,
   selectCurrentRootForm,
   selectPostRecords,
   selectSubRecords,
-} from '../../reducers/recorder';
+} from '../../reducers/Recorder/recorder.selectors';
 
 import { RecordEditorComponent } from './RecordEditor.component';
 
 export const RecordEditorContainer: FC = () => {
   const dispatch = useAppDispatch();
+
+  const { actions } = recorderSlice;
 
   // load data
   useEffect(() => {
@@ -84,7 +88,7 @@ export const RecordEditorContainer: FC = () => {
     (key: string, value: any) => {
       if (currentRecord) {
         dispatch(
-          recorderActions.setFieldValue({
+          actions.setFieldValue({
             recordId: currentRecord.id,
             fieldId: key,
             value,
@@ -113,7 +117,7 @@ export const RecordEditorContainer: FC = () => {
       const subFormId = field.id;
 
       dispatch(
-        recorderActions.addSubRecord({
+        actions.addSubRecord({
           formId: subFormId,
           ownerFieldId,
           ownerRecordId: currentRecord.id,
@@ -138,7 +142,7 @@ export const RecordEditorContainer: FC = () => {
     if (currentRecord.formId !== formIdFromPath) {
       if (currentRecord.ownerId) {
         dispatch(
-          recorderActions.selectRecord({
+          actions.selectRecord({
             recordId: currentRecord.ownerId,
           }),
         );
@@ -151,15 +155,16 @@ export const RecordEditorContainer: FC = () => {
         console.log('RESP', recordResponse);
         dispatch(recordActions.addMany(recordResponse));
         history.push(`/browse/records/${recordResponse[0].id}`);
-      } catch (e) {
+      } catch (e: any) {
         console.log('ERROR', e);
+        dispatch(actions.setRecordErrors({ errors: e?.details?.causes }));
       }
     }
   }, [dispatch, formIdFromPath, currentRecord, recordsToPost, currentForm]);
 
   const handleSelectSubRecord = useCallback(
     (subRecordId: string) => {
-      dispatch(recorderActions.selectRecord({ recordId: subRecordId }));
+      dispatch(actions.selectRecord({ recordId: subRecordId }));
     },
     [dispatch],
   );
@@ -177,6 +182,7 @@ export const RecordEditorContainer: FC = () => {
       onChangeValue={handleFieldValueChange}
       fields={currentForm?.fields}
       values={currentRecord?.values}
+      errors={currentRecord?.errors}
       onAddSubRecord={handleAddSubRecord}
       onSaveRecord={handleSaveRecord}
       subRecords={subRecords}
