+++ frontend/apps/pwa/src/reducers/Former/former.adapter.ts
@@ -1,4 +1,4 @@
-import { createEntityAdapter, createSlice } from '@reduxjs/toolkit';
+import { createEntityAdapter } from '@reduxjs/toolkit';
 
 import { Form } from './types';
