+++ frontend/apps/pwa/src/reducers/__tests__/recorder.spec.ts
@@ -1,4 +1,4 @@
-import { selectPostRecords } from '../recorder';
+import { selectPostRecords } from '../Recorder';
 
 describe('selectPostRecords', () => {
   it('should return the ordered records', function () {
